+++ apps/api/src/services/market-data/market-data.service.ts
@@ -60,12 +60,18 @@ export class MarketDataService {
 
   public async getRange({
     assetProfileIdentifiers,
-    dateQuery
+    dateQuery,
+    skip,
+    take
   }: {
     assetProfileIdentifiers: AssetProfileIdentifier[];
     dateQuery: DateQuery;
+    skip?: number;
+    take?: number;
   }): Promise<MarketData[]> {
     return this.prismaService.marketData.findMany({
+      skip,
+      take,
       orderBy: [
         {
           date: 'asc'
@@ -75,17 +81,33 @@ export class MarketDataService {
         }
       ],
       where: {
-        dataSource: {
-          in: assetProfileIdentifiers.map(({ dataSource }) => {
-            return dataSource;
-          })
-        },
         date: dateQuery,
-        symbol: {
-          in: assetProfileIdentifiers.map(({ symbol }) => {
-            return symbol;
-          })
-        }
+        OR: assetProfileIdentifiers.map(({ dataSource, symbol }) => {
+          return {
+            dataSource,
+            symbol
+          };
+        })
+      }
+    });
+  }
+
+  public async getRangeCount({
+    assetProfileIdentifiers,
+    dateQuery
+  }: {
+    assetProfileIdentifiers: AssetProfileIdentifier[];
+    dateQuery: DateQuery;
+  }): Promise<number> {
+    return this.prismaService.marketData.count({
+      where: {
+        date: dateQuery,
+        OR: assetProfileIdentifiers.map(({ dataSource, symbol }) => {
+          return {
+            dataSource,
+            symbol
+          };
+        })
       }
     });
   }